requirements.txt
torch-ema
ninja
trimesh
opencv-python
tensorboardX
numpy 
pandas
tqdm
matplotlib
PyMCubes
rich
dearpygui
packaging
scipy
scikit-learn

face_alignment
python_speech_features
numba
resampy
soundfile==0.12.1
einops
configargparse

lpips==0.1.3
imageio-ffmpeg
transformers==4.46.2
edge_tts
flask
flask_sockets
opencv-python-headless
aiortc
aiohttp_cors

ffmpeg-python
omegaconf
diffusers
accelerate
librosa
openai
websockets==12.0

funasr
# --- RAG 新增依赖 ---
chromadb # 向量数据库
ocrmypdf # PDF OCR处理
python-docx # 处理 .docx 文件
sentence-transformers # 文本嵌入/向量化
langchain # 用于文本分割等RAG辅助功能

opencv-python-headless # 无头版本的opencv
